+++ lib/hir-mir/src/transformation/equal_operation.rs
@@ -1,3 +1,4 @@
 pub mod expression;
 pub mod function;
 pub mod module;
+mod operation;
